refec: app.js 내부의 이벤트리스너 함수들 각 컴포넌트 내부로 이전
