+++ src/App.js
@@ -10,7 +10,6 @@ import Subtitle from './components/Subtitle.jsx';
 const App = () => {
   const [todos, setTodos] = useState([]);
   const [folded, setFolded] = useState(false);
-  const [value, setValue] = useState('');
   const [todosFetchState, refetch] = useFetch({ fetchUrl: FetchUrl });
   const { data, loading, error } = todosFetchState;
   useEffect(() => {
@@ -18,67 +17,20 @@ const App = () => {
     // eslint-disable-next-line
   }, [data]);
 
-  const onChange = ({ target }) => {
-    setValue(target.value);
-  };
-
-  const onCreate = () => {
-    setTodos([
-      ...todos,
-      {
-        id: Date.now(),
-        title: value,
-        status: 'todo'
-      }
-    ]);
-    setValue('');
-  };
-
-  const onKeyPress = ({ key }) => {
-    if (key === 'Enter') {
-      onCreate();
-    }
-  };
-
-  const onToggle = id => {
-    const index = todos.findIndex(todo => todo.id === id);
-    const selected = todos[index];
-    const selectedStatus = selected.status;
-    const nextTodos = [...todos];
-    nextTodos[index] = {
-      ...selected,
-      status: selectedStatus === 'done' ? 'todo' : 'done'
-    };
-    setTodos(nextTodos);
-  };
-
-  const onRemove = id => {
-    const nextTodos = todos.filter(todo => todo.id !== id);
-    setTodos(nextTodos);
-  };
-
   const onFold = () => {
     setFolded(!folded);
   };
 
   return (
     <TodoListTemplate
-      form={
-        <Form
-          value={value}
-          onCreate={onCreate}
-          onChange={onChange}
-          onKeyPress={onKeyPress}
-        />
-      }
+      form={<Form todos={todos} setTodos={setTodos} />}
       status={<Status todos={todos} />}
       folded={folded}
       subtitle={<Subtitle folded={folded} onFold={onFold} />}
     >
       <TodoItemList
-        onToggle={onToggle}
-        onRemove={onRemove}
         todos={todos}
+        setTodos={setTodos}
         loading={loading}
         error={error}
         refetch={refetch}
